requirements.txt
streamlit
numpy
pandas
scikit-learn
notebook
PyYAML

-e . 